requirements.txt
flask
python-dotenv
pymupdf
chromadb
sentence-transformers
openai
requests
beautifulsoup4
lxml
